Promote Slots version of Foundation from experiments to Foundation package.
