--- packages/experiments/src/utilities/utilities.ts
@@ -1,2 +1,0 @@
-import { __assign } from 'tslib';
-export const assign = __assign;
