+++ packages/foundation/src/IComponent.ts
