+++ packages/foundation/src/ISlots.ts
