+++ packages/foundation/src/index.ts
@@ -1,4 +1,9 @@
 export * from './createComponent';
+export * from './createComponent';
+export * from './IComponent';
+export * from './IHTMLSlots';
+export * from './ISlots';
+export * from './slots';
 export * from './ThemeProvider';
 
 import './version';
